requirements.txt
Flask~=1.1
MarkupSafe==2.0.1
tensorflow
numpy
matplotlib
keras
pillow
urllib3